1.x dev (#3)
* fixes README.md

* adds strict type declarations

* updates phpunit package and fixes test

* adds php 7.3 to travis config

* adds changelog

* creates a dedicated namespace for test classes and move the TestDbEntity class to a Mocks directory as it is a custom mocking class

* Update CHANGELOG.md
