+++ CHANGELOG.md
@@ -1,0 +1,32 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+### Added
+- add changelog
+
+### Changed
+- add strict_types and return type declarations
+- update PHP 7.3 to Travis CI configuration
+- update Travis CI to run tests on PHP 7.3
+- update phpunit/phpunit package to ^7.5
+
+### Fixed
+- fixes headlines and php version in README.md 
+
+## [0.2.0] - 2018-09-18
+### Changed
+- update dependencies to PHP >=7.1 and Symfony ^4.0
+- update starlit inter-package dependencies
+- update Travis CI to run tests on PHP 7.1 and 7.2
+
+## [0.1.1] - 2017-03-26
+### Fixed
+- Fix starlit db requirement
+
+## [0.1.0] - 2017-02-18
+### Added
+- initial release
