+++ README.md
@@ -11,8 +11,8 @@ Add the package as a requirement to your `composer.json`:
 $ composer require starlit/db-data-validation
 ```
 
-##Requirements
-- Requires PHP 7.0 or above.
+## Requirements
+- Requires PHP 7.1 or above.
 
-##License
+## License
 This software is licensed under the BSD 3-Clause License - see the `LICENSE` file for details.
