+++ composer.json
@@ -18,11 +18,14 @@
 
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7",
+        "phpunit/phpunit": "^7.5",
         "symfony/translation": "^4.0"
     },
     "autoload": {
         "psr-4": {"Starlit\\DbDataValidation\\": "src"}
+    },
+    "autoload-dev": {
+        "psr-4": {"Starlit\\DbDataValidation\\Tests\\": "tests"}
     },
     "suggest": {
         "symfony/translation": "Allow translation of validator error messages"
