+++ src/DbEntityDataValidator.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * Db data validation.
  *
